Revise comments. Rescue -> repair
