+++ scripts/change-mode.sh
@@ -19,8 +19,8 @@
 
 # Don't start as root
 if [[ $EUID -eq 0 ]]; then
-   echo "Please start this script als non-root-user, e.g. as user 'pi'" 
-   exit 1
+    echo "Run the script as a regular user"
+    exit 1
 fi
 
 CONFIG_DIR=$(dirname "$0")
@@ -29,10 +29,10 @@ source "$CONFIG"
 
 BACKTITLE="Little Backup Box"
 
-OPTIONS=(1 "Source backup - external storage"
-    2 "Source backup - internal storage"
-    3 "Camera backup - external storage"
-    4 "Camera backup - internal storage")
+OPTIONS=(1 "Source -> external storage"
+    2 "Source -> internal storage"
+    3 "Camera -> external storage"
+    4 "Camera -> internal storage")
 
 CHOICE=$(dialog --clear \
     --backtitle "$BACKTITLE" \
