+++ scripts/lang/en.ini
@@ -2,15 +2,15 @@ sysinfo = "System info"
 status = "Transfer status"
 main = "Main"
 config = "Settings"
-rescue = "Rescue"
+repair = "Repair"
 edit_save_b = "Save"
-backup_storage_external_b = "Source-Backup external"
-backup_storage_external_m = "Backup from one external to another external storage initiated"
-backup_storage_internal_b = "Source-backup internal"
-backup_storage_internal_m = "Backup from external storage to system-storage initiated"
-backup_camera_external_b = "Camera-backup external"
+backup_storage_external_b = "Source (external)"
+backup_storage_external_m = "Backup from source to external storage initiated"
+backup_storage_internal_b = "Source (internal)"
+backup_storage_internal_m = "Backup from source to internal storage initiated"
+backup_camera_external_b = "Camera (external)"
 backup_camera_external_m = "Backup from camera to external storage initiated"
-backup_camera_internal_b = "Camera-backup internal"
+backup_camera_internal_b = "Camera (internal)"
 backup_camera_internal_m = "Backup from camera to internal storage initiated"
 iosbackup_b = "iOS backup"
 iosbackup_m = "iOS backup initiated"
@@ -27,13 +27,13 @@ custom3_m = "Custom action 3 initiated"
 help = "Help"
 help_txt = "
 <dl>
-    <dt>SOURCE-BACKUP EXTERNAL</dt>
-    <dd>Back up a source device (for example, a card reader with a storage card)</dd>
-	<dt>SOURCE-BACKUP INTERNAL</dt>
+    <dt>SOURCE (EXTERNAL)</dt>
+    <dd>Back up a source device (for example, a card reader with a storage card) to an external storage device</dd>
+	<dt>SOURCE (INTERNAL)</dt>
     <dd>Back up a source device to the internal storage</dd>
-    <dt>CAMERA BACKUP EXTERNAL</dt>
+    <dt>CAMERA (EXTERNAL)</dt>
     <dd>Transfer files directly from the connected camera</dd>
-    <dt>CAMERA BACKUP INTERNAL</dt>
+    <dt>CAMERA (INTERNAL)</dt>
     <dd>Transfer files directly from the connected camera to the internal storage</dd>
     <dt>IOS BACKUP</dt>
     <dd>Back up an iOS device to the internal storage</dd>
@@ -51,20 +51,20 @@ log = "Log file"
 log_txt = "Log file not found. Is logging enabled?"
 refresh_b = "Refresh"
 upload = "Upload"
-fsck_check_b = "Filesystem-Check fsck"
-fsck_check_m = "Filesystem-Check started, please wait until finished!\nChecking "
-fsck_autorepair_b = "Filesystem-Autorepair fsck"
-fsck_autorepair_m = "Filesystem-Autorepair started, please wait until finished!\nRepairing "
-rescue_txt="Handle with care! Use this only, if you know what you do!<br>
+fsck_check_b = "Filesystem check"
+fsck_check_m = "Filesystem check started. Wait until finished!\nChecking... "
+fsck_autorepair_b = "Filesystem repair"
+fsck_autorepair_m = "Filesystem repair started. Wait until finished!\nRepairing... "
+repair="Do not use unless you know what you are doing!<br>
 <b>It could damage your storage!</b>
 Follow this steps:
 <ol>
-        <li>Check your power-supply, it must be absolutely safe for at least 10 more minutes.</li>
-        <li>Be sure, no Backup-job is running: The device will be unmounted in this process.</li>
-        <li>Do not klick anywhere else, until the message appeard. Never start a second process as long as another is running!</li>
-        <li>Use Filesystem-Check at first.</li>
-        <li>If errors appear in the log, try the Autorepair.</li>
-        <li>After finishing 'Power off' and unplug the devices before you boot.</li>
+        <li>Check your power supply. Power interruption may damage storage beyond repair.</li>
+        <li>Make sure that no backup job is running.</li>
+        <li>Do not click any buttons until all check and repair operations are completed. Do not start any processes!</li>
+        <li>Run filesystem check first.</li>
+        <li>If you see any errors in the log, run filesystem repair.</li>
+        <li>When the repair operation is finished, power down the machine and unplug the storage devices before boot.</li>
 </ol>
 "
